fix intent extra
intent extra was a double, needs to be int?

+++ app/src/main/java/lf/lookingfor/MapActivity.java
@@ -37,7 +37,7 @@ public class MapActivity extends AppCompatActivity implements OnMapReadyCallback
     DatabaseHandler dbHandler = new DatabaseHandler();
     ArrayList<Event> events = new ArrayList<>();
     HashMap<MarkerOptions, Event> eventMap = new HashMap<>();
-    int mapRadius = 1000;
+    double mapRadius = 1000;
     LocationManager locationManager;
     Location lastKnownLocation;
